requirements.txt
Django==1.11.5
MySQL-python==1.2.5
django-betterforms==1.1.4
django-enumfields==0.9.0
django-mysql==2.1.0
enum34==1.1.6
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pip==9.0.1
pytz==2017.2
setuptools==28.8.0
six==1.10.0
wheel==0.29.0
